pyproject.toml
[project]
name = "kafka"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "confluent-kafka[avro]>=2.10.0",
    "faust-streaming>=0.11.3",
]

[dependency-groups]
dev = [
    "types-confluent-kafka>=1.3.1",
]
